Add admin users and restructure endpoints to support bans
This commit creates admin users, and allows them to ban users, banned
users are unable to make any API calls and are unable to login.

The commit also adds the framework to check for IP bans, it is however
currently not possible to make an IP ban other than by manual database
calls.